==> 04_prepare_data_using_openmeteo_api.ipynb <==
We download the meteo data using openmeteo api, we use many parameters to improve our model

import openmeteo_requests

import pandas as pd
import requests_cache
from retry_requests import retry

# Setup the Open-Meteo API client with cache and retry on error
cache_session = requests_cache.CachedSession('.cache', expire_after = -1)
retry_session = retry(cache_session, retries = 5, backoff_factor = 0.2)
openmeteo = openmeteo_requests.Client(session = retry_session)

# Make sure all required weather variables are listed here
# The order of variables in hourly or daily is important to assign them correctly below
url = "https://archive-api.open-meteo.com/v1/archive"
params = {
	"latitude": 50.768745,
	"longitude": 16.269667,
	"start_date": "2015-01-01",
	"end_date": "2024-01-01",
	"daily": ["temperature_2m_mean", "temperature_2m_max", "temperature_2m_min", "precipitation_sum", "rain_sum", "snowfall_sum", "wind_speed_10m_max", "wind_direction_10m_dominant", "shortwave_radiation_sum", "cape_mean", "cloud_cover_mean", "cloud_cover_max", "cloud_cover_min", "dew_point_2m_mean", "surface_pressure_mean", "surface_pressure_max", "surface_pressure_min", "relative_humidity_2m_mean", "relative_humidity_2m_max", "relative_humidity_2m_min", "wind_speed_10m_mean", "wind_speed_10m_min", "cape_max", "cape_min", "precipitation_hours"],
	"hourly": "boundary_layer_height",
	"wind_speed_unit": "ms",
}
responses = openmeteo.weather_api(url, params=params)

# Process first location. Add a for-loop for multiple locations or weather models
response = responses[0]
print(f"Coordinates: {response.Latitude()}°N {response.Longitude()}°E")
print(f"Elevation: {response.Elevation()} m asl")
print(f"Timezone difference to GMT+0: {response.UtcOffsetSeconds()}s")

# Process hourly data. The order of variables needs to be the same as requested.
hourly = response.Hourly()
hourly_boundary_layer_height = hourly.Variables(0).ValuesAsNumpy()

hourly_data = {"date": pd.date_range(
	start = pd.to_datetime(hourly.Time(), unit = "s", utc = True),
	end = pd.to_datetime(hourly.TimeEnd(), unit = "s", utc = True),
	freq = pd.Timedelta(seconds = hourly.Interval()),
	inclusive = "left"
)}

hourly_data["boundary_layer_height"] = hourly_boundary_layer_height

hourly_dataframe = pd.DataFrame(data = hourly_data)
print("\nHourly data\n", hourly_dataframe)

# Process daily data. The order of variables needs to be the same as requested.
daily = response.Daily()
daily_temperature_2m_mean = daily.Variables(0).ValuesAsNumpy()
daily_temperature_2m_max = daily.Variables(1).ValuesAsNumpy()
daily_temperature_2m_min = daily.Variables(2).ValuesAsNumpy()
daily_precipitation_sum = daily.Variables(3).ValuesAsNumpy()
daily_rain_sum = daily.Variables(4).ValuesAsNumpy()
daily_snowfall_sum = daily.Variables(5).ValuesAsNumpy()
daily_wind_speed_10m_max = daily.Variables(6).ValuesAsNumpy()
daily_wind_direction_10m_dominant = daily.Variables(7).ValuesAsNumpy()
daily_shortwave_radiation_sum = daily.Variables(8).ValuesAsNumpy()
daily_cape_mean = daily.Variables(9).ValuesAsNumpy()
daily_cloud_cover_mean = daily.Variables(10).ValuesAsNumpy()
daily_cloud_cover_max = daily.Variables(11).ValuesAsNumpy()
daily_cloud_cover_min = daily.Variables(12).ValuesAsNumpy()
daily_dew_point_2m_mean = daily.Variables(13).ValuesAsNumpy()
daily_surface_pressure_mean = daily.Variables(14).ValuesAsNumpy()
daily_surface_pressure_max = daily.Variables(15).ValuesAsNumpy()
daily_surface_pressure_min = daily.Variables(16).ValuesAsNumpy()
daily_relative_humidity_2m_mean = daily.Variables(17).ValuesAsNumpy()
daily_relative_humidity_2m_max = daily.Variables(18).ValuesAsNumpy()
daily_relative_humidity_2m_min = daily.Variables(19).ValuesAsNumpy()
daily_wind_speed_10m_mean = daily.Variables(20).ValuesAsNumpy()
daily_wind_speed_10m_min = daily.Variables(21).ValuesAsNumpy()
daily_cape_max = daily.Variables(22).ValuesAsNumpy()
daily_cape_min = daily.Variables(23).ValuesAsNumpy()
daily_precipitation_hours = daily.Variables(24).ValuesAsNumpy()

daily_data = {"date": pd.date_range(
	start = pd.to_datetime(daily.Time(), unit = "s", utc = True),
	end = pd.to_datetime(daily.TimeEnd(), unit = "s", utc = True),
	freq = pd.Timedelta(seconds = daily.Interval()),
	inclusive = "left"
)}

daily_data["temperature_2m_mean"] = daily_temperature_2m_mean
daily_data["temperature_2m_max"] = daily_temperature_2m_max
daily_data["temperature_2m_min"] = daily_temperature_2m_min
daily_data["precipitation_sum"] = daily_precipitation_sum
daily_data["rain_sum"] = daily_rain_sum
daily_data["snowfall_sum"] = daily_snowfall_sum
daily_data["wind_speed_10m_max"] = daily_wind_speed_10m_max
daily_data["wind_direction_10m_dominant"] = daily_wind_direction_10m_dominant
daily_data["shortwave_radiation_sum"] = daily_shortwave_radiation_sum
daily_data["cape_mean"] = daily_cape_mean
daily_data["cloud_cover_mean"] = daily_cloud_cover_mean
daily_data["cloud_cover_max"] = daily_cloud_cover_max
daily_data["cloud_cover_min"] = daily_cloud_cover_min
daily_data["dew_point_2m_mean"] = daily_dew_point_2m_mean
daily_data["surface_pressure_mean"] = daily_surface_pressure_mean
daily_data["surface_pressure_max"] = daily_surface_pressure_max
daily_data["surface_pressure_min"] = daily_surface_pressure_min
daily_data["relative_humidity_2m_mean"] = daily_relative_humidity_2m_mean
daily_data["relative_humidity_2m_max"] = daily_relative_humidity_2m_max
daily_data["relative_humidity_2m_min"] = daily_relative_humidity_2m_min
daily_data["wind_speed_10m_mean"] = daily_wind_speed_10m_mean
daily_data["wind_speed_10m_min"] = daily_wind_speed_10m_min
daily_data["cape_max"] = daily_cape_max
daily_data["cape_min"] = daily_cape_min
daily_data["precipitation_hours"] = daily_precipitation_hours

daily_dataframe = pd.DataFrame(data = daily_data)
print("\nDaily data\n", daily_dataframe)

However, we must merge the hourly and daily data: bounduary layer height is very important parameter and was only available for hourly resolution

hourly_dataframe

daily_dataframe

hourly_dataframe["date"] = pd.to_datetime(hourly_dataframe['date'])
hourly_dataframe = hourly_dataframe.set_index('date')
hourly_dataframe = hourly_dataframe.resample('D').mean()
hourly_dataframe

daily_dataframe['date'] = pd.to_datetime(daily_dataframe['date'])
daily_dataframe = daily_dataframe.set_index('date')

daily_dataframe = daily_dataframe.join(hourly_dataframe)
daily_dataframe

daily_dataframe.index = daily_dataframe.index.tz_convert(None)
ds = daily_dataframe.to_xarray()
ds

ds.to_netcdf("wal_weather.nc")
